# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Certiwise",
        "description": "Api para generar certificados",
        "contact": {
            "email": "contact@academlo.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/{tenant}/events": {
            "get": {
                "tags": [
                    "Event"
                ],
                "summary": "Obtener lista de eventos",
                "operationId": "2ac8a02dcc68fc78b9b96730004c8976",
                "parameters": [
                    {
                        "name": "tenant",
                        "in": "path",
                        "description": "ID del inquilino",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Academlo"
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Número de eventos por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de eventos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Lista de eventos"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                }
            },
            "post": {
                "tags": [
                    "Event"
                ],
                "summary": "Crear un nuevo evento",
                "operationId": "7ff446cf42ad3cfd9adfb7f9d16d951e",
                "parameters": [
                    {
                        "name": "tenant",
                        "in": "path",
                        "description": "ID del inquilino",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Academlo"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Evento de muestra"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Descripción del evento"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Evento creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Evento creado exitosamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                }
            }
        },
        "/api/v1/{tenant}/events/{id}": {
            "get": {
                "tags": [
                    "Event"
                ],
                "summary": "Obtener detalles de un evento",
                "operationId": "fdb340063694bf958e0ad979b921a5c0",
                "parameters": [
                    {
                        "name": "tenant",
                        "in": "path",
                        "description": "ID del inquilino",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Academlo"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del evento",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles del evento",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Detalles del evento"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "Evento no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Event"
                ],
                "summary": "Actualizar un evento existente",
                "operationId": "8ec8054a7afce2f85eaa65a5fa61ec38",
                "parameters": [
                    {
                        "name": "tenant",
                        "in": "path",
                        "description": "ID del inquilino",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Academlo"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del evento",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Evento actualizado"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Descripción actualizada del evento"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Evento actualizado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Evento actualizado exitosamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "Evento no encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Event"
                ],
                "summary": "Eliminar un evento",
                "operationId": "f30c60c5288b4552f6ec136946b5854b",
                "parameters": [
                    {
                        "name": "tenant",
                        "in": "path",
                        "description": "ID del inquilino",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Academlo"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del evento",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Evento eliminado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Evento eliminado exitosamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "Evento no encontrado"
                    }
                }
            }
        }
    }
}